Understanding of Thread Class and Runnable Interface
I got a in depth understanding of Thread Class and Runnable
Interface👌.Understood the nuansese of Task defination and task
managment.It is preferred to use Runnable Interface so that we can
maintain the separation of concerns. Code flexibily and modularity.

Now, the new industry standards are using Runnable Interface and
Executor service instead Thread Class for Thread Management.